refactor: move web3provider to providers directory
