+++ client/src/components/pages/EditArticlePage/index.tsx
@@ -4,7 +4,7 @@ import { Editor as TinyMCEEditor } from "tinymce";
 import { TINY_MCE_API_KEY } from "~/configs/tinymce";
 import { useCallback, useEffect, useState } from "react";
 import { getArticle, updateArticle } from "~/apis/knowtfolio";
-import { useWeb3 } from "~/components/organisms/Web3Provider";
+import { useWeb3 } from "~/components/organisms/providers/Web3Provider";
 
 const EditArticlePage = () => {
   const { articleId } = useParams();
